requirements.txt
# Core Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Machine Learning and Deep Learning
tensorflow>=2.13.0
xgboost>=1.7.0
lightgbm>=4.0.0
statsmodels>=0.14.0

# Database and Data Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.4.0
redis>=4.6.0

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Visualization and Dashboard
plotly>=5.15.0
dash>=2.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Workflow Orchestration
apache-airflow>=2.7.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0
configparser>=6.0.0

# Logging and Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Financial and Risk Libraries
QuantLib-Python>=1.31
pyfolio>=0.9.2
empyrical>=0.5.5

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
python-dateutil>=2.8.0